@@ -1,8 +1,11 @@
-module github.com/bgokden/gofuture
+module github.com/fengyuan-liang/gofuture
 
-go 1.13
+go 1.18
+
+require github.com/stretchr/testify v1.8.2
 
 require (
-	github.com/stretchr/testify v1.4.0
-	golang.org/x/tools v0.0.0-20191109212701-97ad0ed33101 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
